Added some more components and made player movement a lot smoother.

+++ include/components/position.hpp
@@ -6,15 +6,15 @@
 namespace voidgame {
     class position : public component {
         public:
-            position (std::shared_ptr<entity> _ent, const uint32_t _x, const uint32_t _y) : component (_ent, POSITION), x (_x), y (_y) {}
+            position (std::shared_ptr<entity> _ent, const double _x, const double _y) : component (_ent, POSITION), x (_x), y (_y) {}
 
-            uint32_t get_x () { return x; }
-            uint32_t get_y () { return y; }
-            void set_x (const uint32_t _x) { x = _x; }
-            void set_y (const uint32_t _y) { y = _y; }
+            double get_x () { return x; }
+            double get_y () { return y; }
+            void set_x (const double _x) { x = _x; }
+            void set_y (const double _y) { y = _y; }
 
         private:
-            uint32_t x, y;
+            double x, y;
     };
 }
